app/requirements.txt
Django==1.10
django-bootstrap3==7.1.0
django-jquery==3.1.0
django-smtp-ssl==1.0
django-stronghold==0.2.8
djangorestframework==3.4.4
djangorestframework-jwt==1.9.0
httmock==1.2.6
furl==1.0.1
mock==2.0.0
mysqlclient==1.3.9
py-auth0-jwt==0.2.4
py-auth0-jwt-rest==0.1
pycrypto==2.6.1
PyJWT==1.4.2
python-pstore==0.8
raven==6.1.0
requests==2.11.1
